Allow for GNU strip to be used, in some cases.
We get rid of the TARGET_FAKE_STRIP logic that was don't strip at all, just cp.
This can be done by instead doing TARGET_STRIP_MODULE = false.

Add in GNU strip support, which can only be used when PRELINK is off, so
add a check for that.

Turn off TARGET_STRIP_MODULE (which wasn't supported by the tools) for PPC
and MIPS and turn on TARGET_GNU_STRIP_MODULE for both, as PRELINK is not
supported on either.