Fix classic ppc builds to not have isel instructions and always use -mregnames
Current GCC, when built for powerpc.*linux.*spe will default to -misel (e500
specific) regardless of -mcpu being passed.  So we spell out -misel for e500
and -mno-isel for classic.  But due to how crt.*.S are built we need to be
passing these flags in via the CC/CXX definition rather than CFLAGS.