Fix endian in kcm tool
This is *temporary* fix of endian issue in kcm tool. ONLY for the demo.
Although ByteOrder.h is included into kcm.cpp but it can't be
utilized properly. ByteOrder.h needs to keep DEVICE_BYTE_ORDER
defined as LITTLE_ENDIAN even though PPC is big endian because we want to be
able to use apk files built for little endian targets on the PPC.

In short, the endian handling is a complex issue and will be resolved
properly in the future.